traverse pot
